move pages folder into src
