+++ src/pages/_app.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
-import { initializeStore } from '../src/store'
+import { initializeStore } from '../store'
 import { Provider } from 'react-redux'
 import App, { AppContext } from 'next/app'
 import withRedux from 'next-redux-wrapper'
 import { Store } from 'redux'
-import { loadUserFromLocalStorage } from '../src/session/storage'
-import * as SessionStore from '../src/session/redux'
-import AppEffects from '../src/AppEffects'
+import { loadUserFromLocalStorage } from '../session/storage'
+import * as SessionStore from '../session/redux'
+import AppEffects from '../AppEffects'
 import 'react-vis/dist/style.css'
 import Modal from 'react-modal'
-import '../src/ui/setup'
-import Layout from '../src/ui/components/Layout'
+import '../ui/setup'
+import Layout from '../ui/components/Layout'
 
 class MyApp extends App<{ store: Store }> {
   static async getInitialProps({ Component, ctx }: AppContext) {
